changed file names

+++ rules/peakCorrection.smk
@@ -1,31 +1,38 @@
 rule cutAdapt:
 	input:
-		"data/peaks/peaks-{barcode}.txt"
+		"data/03_LCPs_peaks/peaks_{barcode}.txt"
 	output:
-		temp("data/peaks/input-peaks-{barcode}.txt")
+		temp("data/03_LCPs_peaks/input-peaks-{barcode}.txt")
 	conda:
 		"envs/On-rep-seq.yaml"
+	params:
+		porechopped="data/01_porechopped_data"
+		peaks="data/03_LCPs_peaks"
 	shell:
 		"""
 		sed 1d {input} | while read line
 		do
-			P1=$(echo $line | cut -d',' -f 5)
+			P1=$(echo $line | cut -d',' -f 5 )
 			P2=$(echo $line | cut -d',' -f 6)
-			name=$(echo $line | cut -d',' -f 3)
-			cutadapt -m $P1 data/porechopped_2/{wildcards.barcode}.fastq -o data/peaks/{wildcards.barcode}_short_$name.fastq 
-			cutadapt -M $P2 data/peaks/{wildcards.barcode}_short_$name.fastq -o data/peaks/{wildcards.barcode}_$name.fastq
-			echo "{wildcards.barcode}_$name" >> {output}
-			rm data/peaks/{wildcards.barcode}_short_$name.fastq
+			if [ $P2 > 300 ]
+			then
+				name=$(echo $line | cut -d',' -f 3)
+				cutadapt -m $P1 {params.porechopped)/{wildcards.barcode}.fastq -o {params.peaks)/{wildcards.barcode}_short_$name.fastq 
+				cutadapt -M $P2 {params.peaks)/{wildcards.barcode}_short_$name.fastq -o data/peaks/{wildcards.barcode}_$name.fastq
+				echo "{wildcards.barcode}_$name" >> {output}
+				rm data/peaks/{wildcards.barcode}_short_$name.fastq
+			fi
+
 		done	
 		"""
 
 rule correctReads:
 	input:
-		"data/peaks/input-peaks-{barcode}.txt"
+		"data/03_LCPs_peaks/input-peaks-{barcode}.txt"
 	output:
-		temp("data/peaks/fixed_{barcode}.txt")
+		temp("data/03_LCPs_peaks/fixed_{barcode}.txt")
 	params:
-		"data/peaks"
+		"data/03_LCPs_peaks"
 	conda:
 		"envs/canu.yaml"
 	shell:
@@ -40,29 +47,30 @@ rule correctReads:
         		echo "fixed_$line" >> {output}
         	else
             	touch {output}
-        	fi			
+        	fi
+        rm -rf {params}/fixed_$line
 		done
 		"""
 rule vSearch:
 	input:
-		"data/peaks/fixed_{barcode}.txt"	
+		"data/03_LCPs_peaks/fixed_{barcode}.txt"	
 	output:
-		temp("data/peaks/vsearch_fixed_{barcode}.txt")
+		temp("data/03_LCPs_peaks/00_peak_consensus/vsearch_fixed_{barcode}.txt")
 	params:
-		"data/peaks"
+		LCPs="data/03_LCPs_peaks"
+		consensus="data/03_LCPs_peaks/00_peak_consensus"
 	conda:
 		"envs/On-rep-seq.yaml"
 	shell:
 		"""
 		cat {input} | while read line
 		do
-			count=$(grep -c ">" {params}/$line.fa )
+			count=$(grep -c ">" {params.LCPs}/$line.fa )
 			min=$(echo "scale=0 ; $count / 5" | bc )
-			echo $line
-			vsearch --sortbylength {params}/$line.fa --output {params}/sorted_$line.fa
-			vsearch --cluster_fast {params}/sorted_$line.fa -id 0.9  --consout {params}/consensus_$line.fasta -strand both -minsl 0.80 -sizeout -minuniquesize $min
-			vsearch --sortbysize {params}/consensus_$line.fasta --output {params}/vsearch_$line.fa --minsize 50
-			echo "vsearch_$line" >> {output}
-			#remove all non wanted here
+			echo "$line.fa" >> {output}
+			vsearch --sortbylength {params.LCPs}/$line.fa --output {params.LCPs}/sorted_$line.fasta
+			vsearch --cluster_fast {params.LCPs}/sorted_$line.fasta -id 0.9  --consout {params.LCPs}/consensus_$line.fasta -strand both -minsl 0.80 -sizeout -minuniquesize $min
+			vsearch --sortbysize {params.LCPs}/consensus_$line.fasta --output {params.consensus}/$line.fa --minsize 50
+			rm {params.LCPs}/sorted_$line.fa {params.LCPs}/consensus_$line.fasta
 		done
 		"""
